Refactored initializers
